requirements.txt
heatmapz==0.0.4
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
PyYAML==6.0.2
scipy==1.10.0
seaborn==0.11.2
